requirements.txt
flask==2.3.3
flask-sqlalchemy==3.0.5
flask-migrate==4.0.5
flask-login==0.6.2
flask-wtf==1.1.1
flask-mail==0.9.1
gunicorn==21.2.0
python-dotenv==1.0.0
pandas==2.1.1
matplotlib==3.8.0
plotly==5.17.0
yfinance==0.2.31
requests==2.31.0
Werkzeug==2.3.7
openai==0.28.0
fpdf2==2.7.4
newsapi-python==0.2.6
numpy==1.26.0
itsdangerous==2.1.2
pytz==2023.3
stripe==6.1.0
WTForms==3.0.1
email-validator==2.0.0
Jinja2==3.1.2
MarkupSafe==2.1.3
scikit-learn==1.3.0
python-dateutil==2.8.2
beautifulsoup4==4.12.2
feedparser==6.0.10
psycopg2-binary==2.9.7
redis==4.5.1
celery==5.3.0
colorama==0.4.6
psutil
Flask-Limiter>=3.3.1
Flask-Mail>=0.9.1
sentry-sdk[flask]>=1.28.1
python-dotenv>=1.0.0
boto3>=1.28.1  # For S3 backup
schedule>=1.2.0  # For scheduled tasks
pytest>=7.4.0  # For testing
pytest-cov>=4.1.0  # For test coverage
# Advanced Analytics Dependencies
scipy>=1.11.0  # For portfolio optimization and risk calculations
joblib>=1.3.0  # For ML model persistence
ta-lib>=0.4.0  # For technical analysis indicators
cvxpy>=1.4.0  # For advanced optimization problems
